+++ linkerd/protocol/h2/src/main/scala/io/buoyant/linkerd/protocol/h2/HeaderTokenIdentifier.scala
@@ -53,7 +53,7 @@ class HeaderTokenIdentifierConfig extends H2IdentifierConfig {
 }
 
 object HeaderTokenIdentifierConfig {
-  val kind = "io.l5d.headerToken"
+  val kind = "io.l5d.h2.header.token"
 }
 
 class HeaderTokenIdentifierInitializer extends IdentifierInitializer {